expansion_step11

+++ controllers/blogs.js
@@ -1,36 +1,16 @@
 const blogsRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
-
+const middleware = require('../utils/middleware')
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
-const getTokenFrom = request => {
-
-  const authorization = request.get('authorization')
-
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-
-  return null
-}
 
 blogsRouter.get('/', async (request, response) => {
   const  blogs= await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async(request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async(request, response) => {
   const body=request.body
-  const token = getTokenFrom(request)
-
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
-  //const user = await User.findById(body.userId)
+  const user = request.user
 
 
   const blog = new Blog({
@@ -51,14 +31,25 @@ blogsRouter.post('/', async(request, response) => {
 
 
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  console.log('1',user.id)
+  console.log('2',blog.user)
+
+  if(user.id.toString()===blog.user.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    response.status(404).json({ error: 'user not authorized' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const blog = request.body
-
+  if(!request.token) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const updatedBlog = await Blog
     .findByIdAndUpdate(
       request.params.id,
